Use a single if-statement instead of nested if-statements

+++ bleachbit/Special.py
@@ -400,12 +400,11 @@ def delete_ooo_history(path):
     dom1 = xml.dom.minidom.parse(path)
     changed = False
     for node in dom1.getElementsByTagName("node"):
-        if node.hasAttribute("oor:name"):
-            if "History" == node.getAttribute("oor:name"):
-                node.parentNode.removeChild(node)
-                node.unlink()
-                changed = True
-                break
+        if node.hasAttribute("oor:name") and "History" == node.getAttribute("oor:name"):
+            node.parentNode.removeChild(node)
+            node.unlink()
+            changed = True
+            break
     if changed:
         dom1.writexml(open(path, "w", encoding='utf-8'))
